# perf(ivy): improve styling performance