@gkalpak 2019-10-22
Wow, I am quite surprised the VM does not optimize the RegExp literal and creates a new RegExp each time. Are you saying that `const re = /\s+/; ... newValues.split(re)` makes a difference in performance?

Also, ooc, have you tried `/ +/` (instead of `/\s+/`)?