@mhevery 2019-10-22
> Wow, I am quite surprised the VM does not optimize the RegExp literal and creates a new RegExp each time. Are you saying that `const re = /\s+/; ... newValues.split(re)` makes a difference in performance?

1. `RegExp` have internal state so you can't share them.
2. I think the issue is not that we are allocating `RegExp` but rather that `RegExp` is smart, and hence it is not as fast on simples cases as hand tuned. code.