@pkozlowski-opensource 2019-10-23
I've run benchmarks on this PR and here are the results. Before:

```
  [class]=" '1' ": 291.984 us(-43%)
  [class]=" '1 2' ": 428.390 us(-109%)
  [class]=" '1 2 3 4 5 6 7 8 9 0' ": 1698.592 us(-729%)
  class="A B" [class]=" '1' ": 329.344 us(-61%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' ": 1998.899 us(-876%)
  class="A B" [class]=" '1' " [class.foo]="exp": 389.557 us(-90%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' " [class.foo]="exp": 2088.971 us(-920%)
  [element.class]="exp": 204.865 us(0%)
```

After:
```
  [class]=" '1' ": 264.108 us(-29%)
  [class]=" '1 2' ": 399.907 us(-95%)
  [class]=" '1 2 3 4 5 6 7 8 9 0' ": 1674.186 us(-717%)
  class="A B" [class]=" '1' ": 299.102 us(-46%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' ": 1944.948 us(-849%)
  class="A B" [class]=" '1' " [class.foo]="exp": 354.644 us(-73%)
  class="A B" [class]=" '1 2 3 4 5 6 7 8 9 0' " [class.foo]="exp": 2003.856 us(-878%)
  [element.class]="exp": 204.905 us(0%)
```

so there is a clear improvement here (although I was expecting it to be higher, TBH). Those numbers are confirmed by checking the results under profiler. Before:

![before](https://user-images.githubusercontent.com/973550/67379254-dcc89100-f588-11e9-9bad-1b40ffbe2d5d.png)

After:

![after](https://user-images.githubusercontent.com/973550/67379352-0ed9f300-f589-11e9-9ffe-c1e174209057.png)


One can see that `split` disappeared but `splitOnWhitespace` isn't "free". Sadly, the need for split is at the very hart of the algorithm so I'm not sure we can do much about it.